+++ es-core/src/components/DateTimeComponent.cpp
@@ -90,7 +90,7 @@ bool DateTimeComponent::input(InputConfig* config, Input input)
 			}
 
 			//validate day
-			int days_in_month = boost::gregorian::date(new_tm.tm_year + 1900, new_tm.tm_mon + 1, 1).end_of_month().day().as_number();
+			int days_in_month = boost::gregorian::date((unsigned short)new_tm.tm_year + 1900, (unsigned short)new_tm.tm_mon + 1, 1).end_of_month().day().as_number();
 			if(new_tm.tm_mday > days_in_month)
 				new_tm.tm_mday = days_in_month;
